Version 0.2
In this update for our penguin game, we made it so you have a health bar above your penguin character. This way you can see your health as you go so you know if you are about to die! I hope you guys like the new update to the game.